pyproject.toml
[project]
name = "prism-bench"
version = "1.0.0"
description = "PRISM: Pluralistic Reasoning & Identity-Specific Modeling - A Cultural Intelligence benchmark"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
    "google-genai>=1.36.0",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
